pyproject.toml
[tool.poetry]
name = "goit-web-hw-08"
version = "0.1.0"
description = ""
authors = ["R.Anatolii <tv.weplay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
colorama = "^0.4.6"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"
mongoengine = "^0.29.1"
pika = "^1.3.2"
redis = "^5.1.1"
redis-lru = "^0.1.2"
faker = "^30.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
